Fixed plotting issues
Code from previous commit didn't work when trying to do plotting as argument was named plotting (same as nextorch module). Also, as objective function exists, nextorch requires it to be evaluated at every mesh point created. I solved this, but it's not an ideal solution, if only for the impact on the readability of the code. I'd like to submit a PR with a fix for this, although it doesn't seem like nextorch is being actively maintained.